% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsd.R
\name{amsd}
\alias{amsd}
\title{Run Aggragate Mutation Spectrum Distance Method}
\usage{
amsd(
  set1,
  set2,
  mean_or_sum = "mean",
  n_sim = 1000,
  seed = NULL,
  override = FALSE
)
}
\arguments{
\item{set1}{A matrix or data.frame of mutation spectra for group 1}

\item{set2}{A matrix or data.frame of mutation spectra for group 2}

\item{mean_or_sum}{Either "mean" (default) or "sum" to specify whether to
aggregate spectrum is an average of the group's spectra ("mean": when
comparing mutation spectra fractions that add up to 1) or sum ("sum"
when comapring mutation counts)}

\item{n_sim}{Number of null resampling permutations to run (default = 1000)}

\item{seed}{Specify seed for reproducibility (default = NULL)}

\item{override}{Logical; if TRUE, bypasses the rowSum validation checks}
}
\value{
A list of three results: $cosine = observed cosine distance between
the aggragate group spectra, $p  = p-value denoting the fraction of the
random sampling permutations that are greater than or equal to the
observed cosine distance, $sims = cosine distance for each of the random
sampling permutations.
}
\description{
AMSD calculates significance of a difference between two groups of
mutation spectra by comparing the distance between the aggregate spectra
to a null distribution produced by random sampling, outputting a p-value.
}
