% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_dist.R
\name{cosine_dist}
\alias{cosine_dist}
\title{Calculate Cosine Distance}
\usage{
cosine_dist(A, B)
}
\arguments{
\item{A}{A vector (e.g. mutation spectrum: 96 3mer contexts)}

\item{B}{A second vector}
}
\value{
Cosine distance between the vectors (0-1, 0 = identical)
}
\description{
This function calculates the cosine distance between two vectors
}
