% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mouse_carcinogen_counts}
\alias{mouse_carcinogen_counts}
\title{Example mutation count data}
\format{
A data frame with samples as rows and mutation types as columns.
}
\source{
Internal example data.
}
\usage{
mouse_carcinogen_counts
}
\description{
Counts of mutations from mouse tumors exposed to carcinogens.
}
\keyword{datasets}
