% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{mouse_sample_table}
\alias{mouse_sample_table}
\title{Metadata table for mouse tumor samples}
\format{
A data frame with 4 columns: label, tissue, exposure, rep.
}
\usage{
mouse_sample_table
}
\description{
Includes sample labels, tissue types, exposure, and replicate number.
}
\keyword{datasets}
