% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amsd_histogram.R
\name{plot_amsd_histogram}
\alias{plot_amsd_histogram}
\title{Plot Histogram of AMSD Output}
\usage{
plot_amsd_histogram(output)
}
\arguments{
\item{output}{The output from amsd (list of three: $cosine, $p, $sims)}
}
\value{
A ggplot histogram
}
\description{
Generates a ggplot histogram of the AMSD permutations, with the observed
cosine distance marked by a line and the p-value in the title.
}
