% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spectra.R
\name{simulate_spectra}
\alias{simulate_spectra}
\title{Simulates Mutation Spectra for AMSD Input}
\usage{
simulate_spectra(
  n_samples = 1,
  n_mutations = 100,
  sig_probs = c(SBS1 = 0.3, SBS5 = 0.6, SBS18 = 0.1),
  signatures,
  additional_sig = NULL,
  n_extra = 0,
  seed = NULL
)
}
\arguments{
\item{n_samples}{Number of samples to simulate}

\item{n_mutations}{Number of total mutations per sample}

\item{sig_probs}{Specify signatures to sample from and the fraction of
signatures to sample from each}

\item{signatures}{Matrix of signature spectra, with each row a signature
specified by rowname. Must incude names from sig_probs}

\item{additional_sig}{Additional signature to add as an exposure}

\item{n_extra}{Number of mutations to sample from additional signature}

\item{seed}{Specify seed for reproducibility (default = NULL)}
}
\value{
List n_samples long of simluated spectra mutation counts
}
\description{
Generates simulated mutation spectra
}
